requirements.txt
aiogram==3.20.0.post0
Telethon==1.35.0
aiosqlite==0.19.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2